+++ example/lib/constants.dart
@@ -23,4 +23,4 @@ final heroNav = BoxDecoration(
     ]);
 
 
-const token = "";
+const token = "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";