+++ lib/src/model/chat_complete/response/chat_choice.dart
@@ -1,6 +1,7 @@
 import 'message.dart';
 
 class ChatChoice {
+  final String id = "${DateTime.now().millisecondsSinceEpoch}";
   final int index;
   final Message message;
   final String? finishReason;