+++ lib/src/model/complete_text/request/complete_text.dart
@@ -1,6 +1,32 @@
+import 'package:chat_gpt_sdk/src/utils/constants.dart';
+
+enum Model { TextDavinci3, kTextDavinci2, kCodeDavinci2 }
+
+extension ModelExtension on Model {
+  String get name {
+    switch (this) {
+      case Model.TextDavinci3:
+        return kTextDavinci3;
+      case Model.kTextDavinci2:
+        return kTextDavinci2;
+      case Model.kCodeDavinci2:
+        return kCodeDavinci2;
+      default:
+        return "";
+    }
+  }
+}
+
+Model _fromName(String name) {
+  for (var value in Model.values) {
+    if (value.name == name) return value;
+  }
+  throw ArgumentError.value(name, "name", "No enum value with that name");
+}
+
 class CompleteText {
   final String prompt;
-  final String model;
+  final Model model;
   final double temperature;
   final int maxTokens;
   final double topP;
@@ -27,7 +53,7 @@ class CompleteText {
 
   factory CompleteText.fromJson(Map<String, dynamic> json) => CompleteText(
         prompt: json['prompt'] as String,
-        model: json['model'] as String,
+        model: _fromName(json['model'].toString()), //json['model'] as String,
         temperature: (json['temperature'] as num?)?.toDouble() ?? .3,
         maxTokens: json['max_tokens'] as int? ?? 100,
         topP: (json['top_p'] as num?)?.toDouble() ?? 1.0,
@@ -35,17 +61,15 @@ class CompleteText {
         presencePenalty: (json['presence_penalty'] as num?)?.toDouble() ?? .0,
         stop: (json['stop'] as List<String>?),
       );
-  Map<String, dynamic> toJson() => completeReqToJson(this);
 
-  Map<String, dynamic> completeReqToJson(CompleteText instance) =>
-      <String, dynamic>{
-        'prompt': instance.prompt,
-        'model': instance.model,
-        'temperature': instance.temperature,
-        'max_tokens': instance.maxTokens,
-        'top_p': instance.topP,
-        'frequency_penalty': instance.frequencyPenalty,
-        'presence_penalty': instance.presencePenalty,
-        "stop": instance.stop
+  Map<String, dynamic> toJson() => <String, dynamic>{
+        'prompt': prompt,
+        'model': model.name,
+        'temperature': temperature,
+        'max_tokens': maxTokens,
+        'top_p': topP,
+        'frequency_penalty': frequencyPenalty,
+        'presence_penalty': presencePenalty,
+        "stop": stop
       };
 }