+++ lib/src/model/complete_text/response/usage.dart
@@ -2,6 +2,7 @@ class Usage {
   final int promptTokens;
   final int? completionTokens;
   final int totalTokens;
+  final String id = "${DateTime.now().millisecondsSinceEpoch}";
 
   Usage(this.promptTokens, this.completionTokens, this.totalTokens);