+++ lib/src/model/gen_image/response/GenImgResponse.dart
@@ -8,6 +8,7 @@ class GenImgResponse {
 
   int? created;
   List<ImageData?>? data;
+  final String conversionId = "${DateTime.now().millisecondsSinceEpoch}";
 
   factory GenImgResponse.fromJson(Map<String, dynamic> json) => GenImgResponse(
         created: json["created"],